Update the path of files
